+++ include/public/Map.h
@@ -42,6 +42,6 @@ Vector* Map_keys (Map* self);
 
 Vector* Map_values (Map* self);
 
-uint64_t Map_length (Map* self);
+size_t Map_length (Map* self);
 
 hash_t Map_hash (Map* self);
